+++ lib/iterable.rb
@@ -19,6 +19,8 @@ files = %w[
   push_templates
   users
   workflows
+  metadata
+  metadata_table
 ]
 
 files.each { |path| require_relative "./iterable/#{path}" }